Fix FileNotFoundException on Android 13 (API 33+)
READ_MEDIA_IMAGES does not seem to be enough to read jpg files in the audio folders.
Check a file is readable before reading it to avoid a permission-related FileNotFoundException.

This is just a quick fix to remove the exception until a better way to read the cover art files is found (#816).