Fix tag editor on Android R+
Rely on MediaStore.createWriteRequest to get write access to the audio files if the app does not have SD Card permission.

This also moves the permission dialog to the point of save, rather than before any change is done.
Previously, rejecting the permission was leaving the UI in a strange state: all fields were empty.